+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Common & useful classes, resources, extensions. Based on Symfony framework.
 
 1. Resources > translations > add new
 2. Resources > translations > group by destination in each translation domain
+3. Tests > verify default configuration parameters > add "defaults" environment
 
 # 0.1.4